// packages/shared-types/src/entities/UserProfile.d.ts
import { User } from "./User";
export declare class UserProfile {
    id: number;
    user: User;
    name: string;
    grade: number;
    school: string;
    targetExamDate: Date;
    baselineLevel: Record<string, any>;
    currentLevel: Record<string, any>;
    learningPath: Record<string, any>;
}
//# sourceMappingURL=UserProfile.d.ts.map